add restrictions on who can view cases or groups
